Experimental: Use param ids for reordering items
Determining if parameters are "the same" is difficult for pytest:
- some user provided parameters may not be comparable/hashable
- some are, but their equality/hash is determined by identity rather
  than by value, which is likely not a good grouping criterion
- implicit fallbacks are likely to surprise the the user

To solve this, we now rely on the parameter id as it was given
via pytest.param(id=...,) or autogenerated by pytest. This way a user
can explicitely tell pytest which parameters are "the same".

This commit is somewhat larger, becuase we must take care that SubRequests
and reordering agree in their meaning of "the same".